+++ packages/furo-form/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/form@1.3.5...@furo/form@1.3.6) (2020-04-24)
+
+**Note:** Version bump only for package @furo/form
+
+
+
+
+
 ## [1.3.5](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/form@1.3.4...@furo/form@1.3.5) (2020-04-22)
 
 **Note:** Version bump only for package @furo/form