+++ packages/furo-icon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.5](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.4...@furo/icon@1.3.5) (2020-04-24)
+
+**Note:** Version bump only for package @furo/icon
+
+
+
+
+
 ## [1.3.4](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/icon@1.3.3...@furo/icon@1.3.4) (2020-04-21)
 
 **Note:** Version bump only for package @furo/icon